+++ mayan/apps/sources/views.py
@@ -30,7 +30,10 @@ from .exceptions import SourceException
 from .forms import (
     NewDocumentForm, NewVersionForm, WebFormUploadForm, WebFormUploadFormHTML5
 )
-from .icons import icon_log, icon_setup_sources, icon_upload_view_link
+from .icons import (
+    icon_log, icon_setup_sources, icon_staging_folder_file,
+    icon_upload_view_link
+)
 from .literals import SOURCE_UNCOMPRESS_CHOICE_ASK, SOURCE_UNCOMPRESS_CHOICE_Y
 from .links import (
     link_setup_source_create_imap_email, link_setup_source_create_pop3_email,
@@ -155,6 +158,17 @@ class UploadBaseView(MultiFormView):
                         'name': 'appearance/generic_list_subtemplate.html',
                         'context': {
                             'hide_link': True,
+
+                            'no_results_icon': icon_staging_folder_file,
+                            'no_results_text': _(
+                                'This could mean that the staging folder is '
+                                'empty. It could also mean that the '
+                                'operating system user account being used '
+                                'for Mayan EDMS doesn\'t have the necessary '
+                                'file system permissions for the folder.'
+                            ),
+                            'no_results_title': _('No staging files available'),
+
                             'object_list': staging_filelist,
                             'title': _('Files in staging path'),
                         }